use build tags instead of isWindows function
